+++ CHANGELOG.md
@@ -12,7 +12,7 @@ FEATURES:
 * Reset device to Factory Settings
 * Unlock door
 * Get/set basic settings
-* Get advanced settings
+* Get/set advanced settings
 
 BUG FIXES: